feat: Create get user information API
