+++ src/main/java/com/kusitms/backend/controller/UserController.java
@@ -3,7 +3,7 @@ package com.kusitms.backend.controller;
 import com.kusitms.backend.config.SecurityUtil;
 import com.kusitms.backend.dto.UserDto;
 import com.kusitms.backend.response.BaseResponse;
-import com.kusitms.backend.service.IHouseService;
+import com.kusitms.backend.service.IUserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
@@ -14,11 +14,11 @@ import org.springframework.web.bind.annotation.RequestMapping;
 @RequiredArgsConstructor
 @RequestMapping("/api/user")
 public class UserController {
-  private final IHouseService houseService;
+  private final IUserService userService;
 
   @GetMapping
   public ResponseEntity<BaseResponse> getUserInfo() {
-    UserDto response = houseService.getUserInfo(SecurityUtil.getCurrentUserId());
+    UserDto response = userService.getUserInfo(SecurityUtil.getCurrentUserId());
     return ResponseEntity.ok(BaseResponse.builder()
         .data(response).message("유저 정보 조회 성공").build());
   }
