+++ src/main/java/com/kusitms/backend/service/IUserService.java
@@ -1,10 +1,13 @@
 package com.kusitms.backend.service;
 
 import com.kusitms.backend.dto.CheckSmsRequest;
+import com.kusitms.backend.dto.UserDto;
 
 public interface IUserService {
 
   void sendSmsCode(String contact);
 
   void checkSmsCode(CheckSmsRequest request, String email);
+
+  UserDto getUserInfo(String email);
 }
